+++ CHANGELOG.md
@@ -125,6 +125,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [#122](https://github.com/thejoltjoker/kreacon/pull/122)
 - Add Sentry
   [#140](https://github.com/thejoltjoker/kreacon/pull/140)
+- Update ticket api validation for Beacon 2026 event
+  [#153](https://github.com/thejoltjoker/kreacon/pull/153)
 
 ### Deprecated